[IMP] support disabling oca hooks through env var
Closes #115.

As is the case with pylint, this commit adds support for the environment
variable `OCA_HOOKS_DISABLE_CHECKS`. Checks must be listed in it as a
comma separated list, said checks will be added to the disable section
of oca_hooks.cfg, effectively disabling them.

This lets projects disable checks through variables.sh